requirements.txt
discord.py>=1.5.1
SpeechRecognition>=3.8.1
python-dotenv>=0.15.0
pynacl>=1.4.0
